+++ web/src/components/core/index.js
@@ -20,7 +20,6 @@
  */
 
 export { default as About } from "./About";
-export { default as Attribute } from "./Attribute";
 export { default as Description } from "./Description";
 export { default as Disclosure } from "./Disclosure";
 export { default as Sidebar } from "./Sidebar";
@@ -48,4 +47,5 @@ export { default as ProgressReport } from "./ProgressReport";
 export { default as ProgressText } from "./ProgressText";
 export { default as ValidationErrors } from "./ValidationErrors";
 export { default as Terminal } from "./Terminal";
+export { default as Tip } from "./Tip";
 export { default as ShowTerminalButton } from "./ShowTerminalButton";
